* 2.x:
  minor #897 phpstan level 6 fixes (acrobat, GrahamCampbell)
  Added AUTH_ACCESS_TOKEN to allowed methods

+++ lib/Github/HttpClient/Plugin/GithubExceptionThrower.php
@@ -20,7 +20,7 @@ use Psr\Http\Message\ResponseInterface;
 class GithubExceptionThrower implements Plugin
 {
     /**
-     * {@inheritdoc}
+     * @return Promise
      */
     public function handleRequest(RequestInterface $request, callable $next, callable $first): Promise
     {
@@ -32,8 +32,8 @@ class GithubExceptionThrower implements Plugin
             // If error:
             $remaining = ResponseMediator::getHeader($response, 'X-RateLimit-Remaining');
             if (null !== $remaining && 1 > $remaining && 'rate_limit' !== substr($request->getRequestTarget(), 1, 10)) {
-                $limit = ResponseMediator::getHeader($response, 'X-RateLimit-Limit');
-                $reset = ResponseMediator::getHeader($response, 'X-RateLimit-Reset');
+                $limit = (int) ResponseMediator::getHeader($response, 'X-RateLimit-Limit');
+                $reset = (int) ResponseMediator::getHeader($response, 'X-RateLimit-Reset');
 
                 throw new ApiLimitExceedException($limit, $reset);
             }
